extended UserService and LocationService interfaces
