+++ src/com/rest/location/service/LocationService.java
@@ -1,0 +1,17 @@
+package com.rest.location.service;
+
+
+import javax.ws.rs.*;
+import javax.ws.rs.core.*;
+import java.sql.*;
+import com.rest.utils.*;
+
+@Path("/location")  	
+public class LocationService {
+	private DBCon dbcn;
+	private Statement st;
+	
+	
+
+
+}
